Refactor: move extensions classes to the newly created extensions folder
Relates to the Api projects
