+++ Managing/Hotel.Managing.Api/Extensions/ServiceCollectionExtensions.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 
-namespace Hotel.Reporting.Api;
+namespace Hotel.Managing.Api.Extensions;
 
 /// <summary>
 /// Расширения коллекции сервисов
@@ -11,7 +11,7 @@ namespace Hotel.Reporting.Api;
 public static class ServiceCollectionExtensions
 {
     /// <summary>
-    /// Добавить сервисы для рбаоты swagger
+    /// Добавить сервисы для работы swagger
     /// </summary>
     /// <param name="collection">Коллекция сервисов</param>
     /// <returns>Коллекция сервисов</returns>
